rework !color-block
It has been mentioned a few times in chat recently that the current !color-block command is not technically correct. It is sometimes possible to use one side of a block with different parts being different colors. It's also been mentioned that layering colors can bring out more colors in the final print than the number of blocks used.